clip-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
open-clip-torch>=2.24.0
torch>=2.0.0
Pillow>=10.0.0
python-multipart>=0.0.6
